+++ CHANGELOG.md
@@ -5,13 +5,15 @@
 - Added support for Emacs and XEmacs
 - Added support for Zsh
 - Added support for LimeChat
-- Added support for Subversion (thx adamme)
-- Added support for Oh My Zsh (thx adamme)
+- Added support for Subversion (thx @adamme)
+- Added support for Oh My Zsh (thx @adamme)
 - Added support for Ruby and Rails (thx @atipugin)
 - Added support for Pow (thx @atipugin)
 - Added support for Ruby Version (thx @adamstac)
+- Added support for Pentadactyl (thx @alanning)
+- Added support for Vimperator (thx @alanning)
 - Improved Git support (thx @atipugin)
-- Improved Bash support (thx adamme)
+- Improved Bash support (thx @adamme)
 - Doc updates
 
 ## Mackup 0.1